# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "WaveOne API",
        "contact": {
            "name": "Aneep Tandel",
            "url": "https://github.com/aneepct"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/auth/v2/login.{format?}": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns authentication parameters.",
                "operationId": "api.login.index",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "contact",
                        "in": "query",
                        "description": "Email or Contact",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "description": "Client ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "client_secret",
                        "in": "query",
                        "description": "Client Secret",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
    },
...
}